requirements.txt
streamlit>=1.37.0
pytesseract>=0.3.13
Pillow>=10.0.0
opencv-python-headless>=4.10.0
numpy>=1.26.0
sympy>=1.12
easyocr>=1.7.0
torch>=2.2.0
torchvision>=0.17.0
